refactor: rename Python package lingoanki → lingodiary
- git mv lingoanki/ lingodiary/ and test_lingoanki/ test_lingodiary/
- pyproject.toml: name, packages include, and 3 script entry-points
- All from-imports inside the package updated (webapp.py, diary.py,
  migrate_to_json.py, audio_timing.py)
- migration/*.sh: python -m lingoanki.xxx → python -m lingodiary.xxx
- release.yml: artifact names lingoanki-wheel/apk → lingodiary-wheel/apk
- Dockerfile: COPY path and comment updated
- .po translation source location comments updated
