+++ .github/workflows/release.yml
@@ -113,7 +113,7 @@ jobs:
       - name: Upload wheel artifact
         uses: actions/upload-artifact@v4
         with:
-          name: lingoanki-wheel
+          name: lingodiary-wheel
           path: dist/*.whl
 
   build-apk:
@@ -174,7 +174,7 @@ jobs:
       - name: Upload APK artifact
         uses: actions/upload-artifact@v4
         with:
-          name: lingoanki-apk
+          name: lingodiary-apk
           path: android_app/build/app/outputs/flutter-apk/app-release.apk
 
   release:
@@ -185,13 +185,13 @@ jobs:
       - name: Download wheel artifact
         uses: actions/download-artifact@v4
         with:
-          name: lingoanki-wheel
+          name: lingodiary-wheel
           path: dist/
 
       - name: Download APK artifact
         uses: actions/download-artifact@v4
         with:
-          name: lingoanki-apk
+          name: lingodiary-apk
           path: apk/
 
       - name: Create GitHub Release
