+++ lingodiary/templates/diary_html.html
