+++ lingodiary/translations/fr/LC_MESSAGES/messages.mo
