+++ lingodiary/translations/fr/LC_MESSAGES/messages.po
@@ -18,194 +18,194 @@ msgstr ""
 "Content-Transfer-Encoding: 8bit\n"
 "Generated-By: Babel 2.17.0\n"
 
-#: lingoanki/templates/diary.html:6
+#: lingodiary/templates/diary.html:6
 msgid "LingoDiary WebApp"
 msgstr "Application Web LingoDiary"
 
-#: lingoanki/templates/diary.html:130
+#: lingodiary/templates/diary.html:130
 msgid "Logged in as"
 msgstr "Connecté en tant que"
 
-#: lingoanki/templates/diary.html:132
+#: lingodiary/templates/diary.html:132
 msgid "Logout"
 msgstr "Se déconnecter"
 
-#: lingoanki/templates/diary.html:138
+#: lingodiary/templates/diary.html:138
 msgid "Step 1: Edit Diary"
 msgstr "Étape 1 : Modifier le journal"
 
-#: lingoanki/templates/diary.html:139
+#: lingodiary/templates/diary.html:139
 msgid "View Diary HTML"
 msgstr "Voir le HTML du journal"
 
-#: lingoanki/templates/diary.html:140
+#: lingodiary/templates/diary.html:140
 msgid "View TPRS"
 msgstr "Voir TPRS"
 
-#: lingoanki/templates/diary.html:141
+#: lingodiary/templates/diary.html:141
 msgid "Step 2: Generate Lessons"
 msgstr "Étape 2 : Générer les leçons"
 
-#: lingoanki/templates/diary.html:142
+#: lingodiary/templates/diary.html:142
 msgid "Step 3: Listen to Lessons"
 msgstr "Étape 3 : Écouter les leçons"
 
-#: lingoanki/templates/diary.html:143
+#: lingodiary/templates/diary.html:143
 msgid "Step 4: Download Output Files"
 msgstr "Étape 4 : Télécharger les fichiers de sortie"
 
-#: lingoanki/templates/diary.html:156
+#: lingodiary/templates/diary.html:156
 msgid "How To Edit Diary"
 msgstr "Comment modifier le journal"
 
-#: lingoanki/templates/diary.html:157
+#: lingodiary/templates/diary.html:157
 msgid "You can edit the diary in two ways:"
 msgstr "Vous pouvez modifier le journal de deux manières :"
 
-#: lingoanki/templates/diary.html:159
+#: lingodiary/templates/diary.html:159
 msgid ""
 "Option 1: By adding sentences directly in the text box below with the "
 "following template and clicking \"Save\"."
 msgstr "Option 1 : En ajoutant des phrases directement dans la zone de texte ci-dessous avec le modèle suivant et en cliquant sur « Enregistrer »."
 
-#: lingoanki/templates/diary.html:160
+#: lingodiary/templates/diary.html:160
 msgid ""
 "Please note that Only the sentence in between the 2 asterixes ** ... ** "
 "needs to be replaced..."
 msgstr "Veuillez noter que seule la phrase entre les deux astérisques ** ... ** doit être remplacée..."
 
-#: lingoanki/templates/diary.html:163
+#: lingodiary/templates/diary.html:163
 msgid ""
 "Option 2: By selecting a specific date from the calendar widget below (or"
 " entering the date manually), then adding sentences for that day."
 msgstr "Option 2 : En sélectionnant une date spécifique dans le calendrier ci-dessous (ou en entrant la date manuellement), puis en ajoutant des phrases pour ce jour-là."
 
-#: lingoanki/templates/diary.html:166
+#: lingodiary/templates/diary.html:166
 msgid "Edit Diary"
 msgstr "Modifier le journal"
 
-#: lingoanki/templates/diary.html:167
+#: lingodiary/templates/diary.html:167
 msgid "Careful not to break the template"
 msgstr "Attention à ne pas casser le modèle"
 
-#: lingoanki/templates/diary.html:170
+#: lingodiary/templates/diary.html:170
 msgid "Save"
 msgstr "Enregistrer"
 
-#: lingoanki/templates/diary.html:174
+#: lingodiary/templates/diary.html:174
 msgid "Option 1"
 msgstr "Option 1"
 
-#: lingoanki/templates/diary.html:175
+#: lingodiary/templates/diary.html:175
 msgid ""
 "Copy template below to the top of the diary and add as many new sentences"
 " as desired"
 msgstr "Copiez le modèle ci-dessous en haut du journal et ajoutez autant de nouvelles phrases que souhaité"
 
-#: lingoanki/templates/diary.html:176
+#: lingodiary/templates/diary.html:176
 msgid "This template has only 3 sentences, but you can add more"
 msgstr "Ce modèle contient seulement 3 phrases, mais vous pouvez en ajouter davantage"
 
-#: lingoanki/templates/diary.html:177
+#: lingodiary/templates/diary.html:177
 msgid "Copy Template"
 msgstr "Copier le modèle"
 
-#: lingoanki/templates/diary.html:190
+#: lingodiary/templates/diary.html:190
 msgid "Option 2"
 msgstr "Option 2"
 
-#: lingoanki/templates/diary.html:191
+#: lingodiary/templates/diary.html:191
 msgid "Select and set a date."
 msgstr "Sélectionnez et définissez une date."
 
-#: lingoanki/templates/diary.html:192
+#: lingodiary/templates/diary.html:192
 msgid "Add as many sentences as wanted and click save"
 msgstr "Ajoutez autant de phrases que vous le souhaitez et cliquez sur enregistrer"
 
-#: lingoanki/templates/diary.html:194
+#: lingodiary/templates/diary.html:194
 msgid "Select a date:"
 msgstr "Sélectionner une date :"
 
-#: lingoanki/templates/diary.html:196
+#: lingodiary/templates/diary.html:196
 msgid "1) Set Date (click me)"
 msgstr "1) Définir la date (cliquez ici)"
 
-#: lingoanki/templates/diary.html:203
+#: lingodiary/templates/diary.html:203
 msgid "Add a new sentence"
 msgstr "Ajouter une nouvelle phrase"
 
-#: lingoanki/templates/diary.html:204
+#: lingodiary/templates/diary.html:204
 msgid "2) Add Sentence"
 msgstr "2) Ajouter une phrase"
 
-#: lingoanki/templates/diary.html:226
+#: lingodiary/templates/diary.html:226
 msgid "Sentence for"
 msgstr "Phrase pour"
 
-#: lingoanki/templates/diary.html:279
+#: lingodiary/templates/diary.html:279
 msgid "Apply"
 msgstr "Appliquer"
 
-#: lingoanki/templates/diary.html:290
+#: lingodiary/templates/diary.html:290
 msgid "3) Save Diary Entry for Date"
 msgstr "3) Enregistrer l'entrée du journal pour la date"
 
-#: lingoanki/templates/diary.html:293
+#: lingodiary/templates/diary.html:293
 msgid "Cancel - Clear Selected Date and Entries"
 msgstr "Annuler - Effacer la date et les entrées sélectionnées"
 
-#: lingoanki/templates/diary_generate_lessons.html:2
+#: lingodiary/templates/diary_generate_lessons.html:2
 msgid "Generate Lessons"
 msgstr "Générer les leçons"
 
-#: lingoanki/templates/diary_generate_lessons.html:4
+#: lingodiary/templates/diary_generate_lessons.html:4
 msgid "Generate"
 msgstr "Générer"
 
-#: lingoanki/templates/diary_html.html:2
+#: lingodiary/templates/diary_html.html:2
 msgid "Diary HTML"
 msgstr "HTML du journal"
 
-#: lingoanki/templates/diary_output.html:2
+#: lingodiary/templates/diary_output.html:2
 msgid "Output Files"
 msgstr "Fichiers de sortie"
 
-#: lingoanki/templates/diary_output.html:9
+#: lingodiary/templates/diary_output.html:9
 msgid "Download All as Zip"
 msgstr "Tout télécharger au format Zip"
 
-#: lingoanki/templates/diary_tprs.html:2
+#: lingodiary/templates/diary_tprs.html:2
 msgid "TPRS Markdown"
 msgstr "Markdown TPRS"
 
-#: lingoanki/templates/diary_tprs.html:5
+#: lingodiary/templates/diary_tprs.html:5
 msgid "Refresh TPRS"
 msgstr "Rafraîchir TPRS"
 
-#: lingoanki/templates/diary_tprs_play_audio.html:5
+#: lingodiary/templates/diary_tprs_play_audio.html:5
 msgid "Select an MP3 file to play"
 msgstr "Sélectionnez un fichier MP3 à lire"
 
-#: lingoanki/templates/diary_tprs_play_audio.html:8
+#: lingodiary/templates/diary_tprs_play_audio.html:8
 msgid "Choose an audio file"
 msgstr "Choisissez un fichier audio"
 
-#: lingoanki/templates/diary_tprs_play_audio.html:14
+#: lingodiary/templates/diary_tprs_play_audio.html:14
 msgid "Choose File"
 msgstr "Choisir un fichier"
 
-#: lingoanki/templates/diary_tprs_play_audio.html:25
+#: lingodiary/templates/diary_tprs_play_audio.html:25
 msgid "Now Playing"
 msgstr "Lecture en cours"
 
-#: lingoanki/templates/diary_tprs_play_audio.html:42
+#: lingodiary/templates/diary_tprs_play_audio.html:42
 msgid "Download this MP3"
 msgstr "Télécharger ce MP3"
 
-#: lingoanki/templates/diary_tprs_play_audio.html:53
+#: lingodiary/templates/diary_tprs_play_audio.html:53
 msgid "Download Markdown"
 msgstr "Télécharger le Markdown"
 
-#: lingoanki/templates/diary_tprs_play_audio.html:60
+#: lingodiary/templates/diary_tprs_play_audio.html:60
 msgid "Markdown Content"
 msgstr "Contenu Markdown"
