+++ migration/backfill_qa_translations.sh
@@ -10,7 +10,7 @@ for USER in claudine johanne laurent; do
   docker compose exec lingo-diary python -c "
 import logging
 logging.basicConfig(level=logging.INFO, format='%(asctime)s %(levelname)s %(message)s')
-from lingoanki.diary import DiaryHandler, _backfill_qa_translations
+from lingodiary.diary import DiaryHandler, _backfill_qa_translations
 handler = DiaryHandler.__new__(DiaryHandler)
 handler.config = DiaryHandler.load_config(handler, '/app/.config/lingoDiary/${USER}/config.yaml')
 logger = logging.getLogger('backfill')
