+++ migration/backfill_sentence_inputs.sh
@@ -13,7 +13,7 @@ for USER in claudine johanne laurent; do
   docker compose exec lingo-diary python -c "
 import logging
 logging.basicConfig(level=logging.INFO, format='%(asctime)s %(levelname)s %(message)s')
-from lingoanki.diary import DiaryHandler, _backfill_variant_sentence_inputs
+from lingodiary.diary import DiaryHandler, _backfill_variant_sentence_inputs
 handler = DiaryHandler.__new__(DiaryHandler)
 handler.config = DiaryHandler.load_config(handler, '/app/.config/lingoDiary/${USER}/config.yaml')
 logger = logging.getLogger('backfill')
