+++ migration/convert_md_json.sh
@@ -2,7 +2,7 @@
 set -e
 cd "$(dirname "$0")/.."
 for USER in claudine johanne laurent; do
-  docker compose exec lingo-diary python -m lingoanki.migrate_to_json \
+  docker compose exec lingo-diary python -m lingodiary.migrate_to_json \
     --config /app/.config/lingoDiary/${USER}/config.yaml \
     --output /data/${USER}/diary.json \
     --overwrite
