+++ pyproject.toml
@@ -1,12 +1,12 @@
 [tool.poetry]
-name = "lingoanki"
+name = "lingodiary"
 version = "2.1.0"
 description = "A script to create Anki flashcards from audio transcript to learn a foreign language."
 authors = ["Your Name <your.email@example.com>"]
 license = "MIT"
 readme = "README.md"
 package-mode = true
-packages = [{ include = "lingoanki" }]
+packages = [{ include = "lingodiary" }]
 
 [tool.poetry.dependencies]
 flask-babel = ">=4.0.0"
@@ -33,9 +33,9 @@ PyJWT = ">=2.8.0"
 
 
 [tool.poetry.scripts]
-lingoAnki = "lingoanki.__main__:main"
-lingoDiary = "lingoanki.diary:main"
-lingoWebapp = "lingoanki.webapp:main"
+lingoAnki = "lingodiary.__main__:main"
+lingoDiary = "lingodiary.diary:main"
+lingoWebapp = "lingodiary.webapp:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
